removed overflow class from #priority element, modified column sizes for #helpInfo and save/load dropdown.
overflow class was no longer needed for #priority element. Changed column sizes for the #helpInfo and save/load dropdown. The dropdown container did not have to be so large as there's little content within it. Whereas the #helpInfo element has alot of content and needed more room, especially for smaller screens.